Simplified the memoization.
After poking around the resolveCtx function's logic a bit more, I
realized that even originalName needn't be used in the memoization key
and only the context is needed. I'm now confident of the correctness
of this optimization and am willing to change the name from "Potential
fix for issue #232" to "Fix for issue #232".

It is easier to see the correctness of the assumption made by the
memoization step in a purely recursive rewrite of resolveCtx, where
every recursive step works with either a "deeper" context or a new one.
I've put that up at https://gist.github.com/srikumarks/9847260 .