Only resolve when absolutely necessary
Many calls to `resolve` are not necessary. Most of the time we only want
to resolve when something is a macro, but if there's no possibility of
it being a macro, we can avoid an expensive call to resolve.

This PR adds an additional map to the env that stores unresolved names
of the things in the env. That way, we can check if it *might* be
a macro, and only then call `resolve`.